requirements.txt
PyPolkadot==0.1.3
git+https://github.com/KarimJedda/swarm.git